@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -39,15 +41,15 @@
 		"bluebird"
 	],
 	"dependencies": {
-		"aggregate-error": "^3.0.0"
+		"aggregate-error": "^4.0.0"
 	},
 	"devDependencies": {
-		"ava": "^2.2.0",
-		"delay": "^4.1.0",
-		"in-range": "^2.0.0",
-		"random-int": "^2.0.0",
-		"time-span": "^3.1.0",
-		"tsd": "^0.7.4",
-		"xo": "^0.27.2"
+		"ava": "^3.15.0",
+		"delay": "^5.0.0",
+		"in-range": "^3.0.0",
+		"random-int": "^3.0.0",
+		"time-span": "^5.0.0",
+		"tsd": "^0.14.0",
+		"xo": "^0.38.2"
 	}
 }
